requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.6
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg==3.1.19
psycopg-binary==3.1.19
PyJWT==2.8.0
pymemcache==4.0.0
python-decouple==3.8
python-memcached==1.62
PyYAML==6.0.1
redis==4.6.0
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
